+++ Cargo.toml
@@ -1,15 +1,8 @@
-[package]
-name = "svart"
-version = "4.2.0"
-edition = "2021"
-
-[dependencies]
-cozy-chess = "0.3.2"
-once_cell = "1.17.1"
-chrono = "0.4.24"
-fastrand = "1.9.0"
-ctrlc = "3.2.5"
-tabled = "0.12.0"
+[workspace]
+members = [
+    "engine",
+    "datagen"
+]
 
 [profile.release]
 lto = "fat"
