+++ datagen/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "datagen"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+cozy-chess = "0.3.2"
+chrono = "0.4.24"
+fastrand = "1.9.0"
+ctrlc = "3.2.5"
+tabled = "0.12.0"
+engine = { path = "../engine" }
