+++ engine/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "engine"
+version = "0.1.0"
+edition = "2021"
+default-run = "engine"
+
+[dependencies]
+cozy-chess = "0.3.2"
+once_cell = "1.17.1"
