+++ engine/src/search/history.rs
@@ -8,9 +8,7 @@ pub struct History {
 
 impl History {
     pub fn new() -> History {
-        History {
-            table: [[[0; 64]; 64]; 2],
-        }
+        History { table: [[[0; 64]; 64]; 2] }
     }
 
     #[must_use]
@@ -40,10 +38,12 @@ impl History {
     }
 
     pub fn age_table(&mut self) {
-        self.table
-            .iter_mut()
-            .flatten()
-            .flatten()
-            .for_each(|x| *x /= 2);
+        self.table.iter_mut().flatten().flatten().for_each(|x| *x /= 2);
+    }
+}
+
+impl Default for History {
+    fn default() -> Self {
+        Self::new()
     }
 }
