+++ engine/src/search/nnue/mod.rs
@@ -1,0 +1,1 @@
+pub mod inference;
