+++ engine/src/search/stat_vec.rs
