--- src/engine/nnue/datagen/mod.rs
@@ -1,1 +1,0 @@
-pub mod script;
