--- src/engine/nnue/mod.rs
@@ -1,3 +1,0 @@
-pub mod datagen;
-pub mod inference;
-pub mod tables;
